Services/requirements.txt
Flask
grpcio
grpcio-tools
pika